pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-ocr-transcript-extractor"
version = "1.0.0"
description = "Extract and deduplicate scrolling text from screen recordings using OCR"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Video OCR Team"}
]
keywords = ["ocr", "video", "transcript", "text-extraction", "teams", "meeting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "opencv-python-headless>=4.8.0",
    "easyocr>=1.7.0",
    "rapidfuzz>=3.0.0",
    "python-Levenshtein>=0.21.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
paddleocr = [
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
]
tesseract = [
    "pytesseract>=0.3.10",
]
web = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
all = [
    "video-ocr-transcript-extractor[paddleocr,tesseract,web,dev]",
]

[project.scripts]
video-ocr = "video_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/video-ocr/video-ocr-transcript-extractor"
Documentation = "https://github.com/video-ocr/video-ocr-transcript-extractor#readme"
Repository = "https://github.com/video-ocr/video-ocr-transcript-extractor"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=video_ocr --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
